+++ src/views/LiveFeed.vue
@@ -1,10 +1,13 @@
 <template>
 	<div>
+		<div v-if="!isAuthenticated">
+			<Login />
+		</div>
 		<my-header />
 		<!-- <div v-if="!isAuthenticated">
 			<Login />
 		</div> -->
-		<div>
+		<div v-if="isAuthenticated">
 			<div class="page-container container">
 				<div class="row">
 					<div class="col-sm-3 mt-3 d-none d-sm-block">
@@ -34,7 +37,7 @@
 											<!-- This returns the number of followers. It's being recieved
 										from the getters in userAuthentication store -->
 											<span v-if="!numberOfFollowers">O</span>
-											<span v-else>{{ numberOfFollowers }}k</span>
+											<span v-else>{{ numberOfFollowers }}</span>
 										</h3>
 										<h4 class="font-11">Followers</h4>
 									</div>
@@ -44,7 +47,7 @@
 											<!-- This returns the number of followers. It's being recieved
 										from the getters in userAuthentication store -->
 											<span v-if="!numberOfFollowed">O</span>
-											<span v-else>{{ numberOfFollowed }}k</span>
+											<span v-else>{{ numberOfFollowed }}</span>
 										</h3>
 										<h4 class="font-11">Following</h4>
 									</div>
@@ -97,6 +100,7 @@
 
 					<div class="col-md-6">
 						<div id="content_area" class="container mt-3">
+							<!-- <PollForming /> -->
 							<PollForm />
 							<div class="card rounded">
 								<!-- Nav tabs -->
@@ -125,6 +129,7 @@
 									<div class="tab-pane active" id="home" role="tabpanel">
 										<div class="card-body">
 											<div class="profiletimeline">
+												<!-- Here we are looping through the allPolls which we received from our getters -->
 												<div
 													class="todo-item sl-right"
 													v-for="poll in allPolls"
@@ -134,7 +139,7 @@
 														<PollMenu></PollMenu>
 														<div class="sl-left">
 															<img
-																src="http://wrappixel.com/demos/admin-templates/admin-wrap/assets/images/users/2.jpg"
+																src="./../assets/img/01_UD.png"
 																alt="user"
 																class="img-circle"
 															/>
@@ -154,7 +159,33 @@
 																	<p>{{ poll.question }}</p>
 																</div>
 															</div>
+															<hr />
+															<div class="m-t-20">
+																<div class="row">
+																	<!-- Here we are looping through the choice array in our poll -->
+																	<div
+																		class=""
+																		v-for="choice in poll.choices"
+																		v-bind:key="choice.id"
+																	>
+																		<div
+																			class="col-md-12 linkHover"
+																			data-toggle="tooltip"
+																			title="vote"
+																			v-if="choice.choice_text !== null"
+																		>
+																			<!-- Here we call the voteChoice method and pass in the poll object and the selected choice id -->
+																			<p @click="voteChoice(poll, choice.id)">
+																				{{ choice.choice_vote_count }}-{{
+																					choice.choice_text
+																				}}
+																			</p>
+																		</div>
+																	</div>
+																</div>
+															</div>
 															<div class="like-comm m-t-20">
+																<!-- Here we call the bookmarkPoll method and pass in the poll object and current user id which we get from our getUser from getters -->
 																<span
 																	class="linkHover m-r-10"
 																	@click="
@@ -162,8 +193,9 @@
 																	"
 																>
 																	<i class="fa fa-thumb text-danger"></i
-																	>{{ poll.vote_count }} Vote</span
+																	>{{ poll.vote_count }} bookmark</span
 																>
+																<!--  Here we call the likePoll method and pass in the poll object and current user id which we get from our getUser from getters -->
 																<span
 																	class="linkHover m-r-10"
 																	@click="
@@ -182,116 +214,7 @@
 									</div>
 									<!--second tab-->
 									<div class="tab-pane" id="profile" role="tabpanel">
-										<div class="card-body">
-											<div class="row">
-												<div class="col-md-3 col-xs-6 b-r">
-													<strong>Full Name</strong>
-													<br />
-													<p class="text-muted">Johnathan Deo</p>
-												</div>
-												<div class="col-md-3 col-xs-6 b-r">
-													<strong>Mobile</strong>
-													<br />
-													<p class="text-muted">(123) 456 7890</p>
-												</div>
-												<div class="col-md-3 col-xs-6 b-r">
-													<strong>Email</strong>
-													<br />
-													<p class="text-muted">johnathan@admin.com</p>
-												</div>
-												<div class="col-md-3 col-xs-6">
-													<strong>Location</strong>
-													<br />
-													<p class="text-muted">London</p>
-												</div>
-											</div>
-											<hr />
-											<p class="m-t-30">
-												Donec pede justo, fringilla vel, aliquet nec, vulputate
-												eget, arcu. In enim justo, rhoncus ut, imperdiet a,
-												venenatis vitae, justo. Nullam dictum felis eu pede
-												mollis pretium. Integer tincidunt.Cras dapibus. Vivamus
-												elementum semper nisi. Aenean vulputate eleifend tellus.
-												Aenean leo ligula, porttitor eu, consequat vitae,
-												eleifend ac, enim.
-											</p>
-											<p>
-												Lorem Ipsum is simply dummy text of the printing and
-												typesetting industry. Lorem Ipsum has been the
-												industry's standard dummy text ever since the 1500s,
-												when an unknown printer took a galley of type and
-												scrambled it to make a type specimen book. It has
-												survived not only five centuries
-											</p>
-											<p>
-												It was popularised in the 1960s with the release of
-												Letraset sheets containing Lorem Ipsum passages, and
-												more recently with desktop publishing software like
-												Aldus PageMaker including versions of Lorem Ipsum.
-											</p>
-											<h4 class="font-medium m-t-30">Skill Set</h4>
-											<hr />
-											<h5 class="m-t-30">
-												Wordpress <span class="pull-right">80%</span>
-											</h5>
-											<div class="progress">
-												<div
-													class="progress-bar bg-success"
-													role="progressbar"
-													aria-valuenow="80"
-													aria-valuemin="0"
-													aria-valuemax="100"
-													style="width:80%; height:6px;"
-												>
-													<span class="sr-only">50% Complete</span>
-												</div>
-											</div>
-											<h5 class="m-t-30">
-												HTML 5 <span class="pull-right">90%</span>
-											</h5>
-											<div class="progress">
-												<div
-													class="progress-bar bg-info"
-													role="progressbar"
-													aria-valuenow="90"
-													aria-valuemin="0"
-													aria-valuemax="100"
-													style="width:90%; height:6px;"
-												>
-													<span class="sr-only">50% Complete</span>
-												</div>
-											</div>
-											<h5 class="m-t-30">
-												jQuery <span class="pull-right">50%</span>
-											</h5>
-											<div class="progress">
-												<div
-													class="progress-bar bg-danger"
-													role="progressbar"
-													aria-valuenow="50"
-													aria-valuemin="0"
-													aria-valuemax="100"
-													style="width:50%; height:6px;"
-												>
-													<span class="sr-only">50% Complete</span>
-												</div>
-											</div>
-											<h5 class="m-t-30">
-												Photoshop <span class="pull-right">70%</span>
-											</h5>
-											<div class="progress">
-												<div
-													class="progress-bar bg-warning"
-													role="progressbar"
-													aria-valuenow="70"
-													aria-valuemin="0"
-													aria-valuemax="100"
-													style="width:70%; height:6px;"
-												>
-													<span class="sr-only">50% Complete</span>
-												</div>
-											</div>
-										</div>
+										<TrendingPolls />
 									</div>
 								</div>
 							</div>
@@ -302,33 +225,7 @@
 						<div class="card rounded">
 							<div class="card-body">
 								<h4 class="card-title">Top Trending</h4>
-								<ul class="feeds">
-									<li>
-										<div class="bg-light-info">
-											<i class="fa fa-bell-o"></i>
-										</div>
-										You have 4 pending tasks.
-										<span class="text-muted">Just Now</span>
-									</li>
-									<li>
-										<div class="bg-light-success">
-											<i class="ti-server"></i>
-										</div>
-										Server #1 overloaded.<span class="text-muted"
-											>2 Hours ago</span
-										>
-									</li>
-									<li>
-										<div class="bg-light-warning">
-											<i class="ti-shopping-cart"></i>
-										</div>
-										New order received.<span class="text-muted">31 May</span>
-									</li>
-									<li>
-										<div class="bg-light-danger"><i class="ti-user"></i></div>
-										New user registered.<span class="text-muted">30 May</span>
-									</li>
-								</ul>
+								<TrendingPolls />
 							</div>
 						</div>
 						<div class="card rounded">
@@ -441,17 +338,30 @@ import PollMenu from "./../components/PollMenu.vue";
 import LoginForm from "./../components/LoginForm.vue";
 import Login from "./Login";
 import PollForm from "@/components/PollForm.vue";
+import PollForming from "./PollForming.vue";
+import TrendingPolls from "./TrendingPolls.vue";
 import { mapGetters, mapActions } from "vuex";
 
 const api = "http://hn.algolia.com/api/v1/search_by_date?tags=story";
 
 export default {
 	name: "feed",
+	components: {
+		PollForming,
+		TrendingPolls,
+		PollForm,
+		LoginForm,
+		Login,
+		PollMenu,
+		InfiniteLoading
+	},
 	data() {
 		return {
 			list: [],
 			selected_poll: null,
-			poll_creator: null
+			poll_creator: null,
+			selected_choice: null,
+			current_date: ""
 		};
 	},
 	methods: {
@@ -474,9 +384,15 @@ export default {
 					}
 				});
 		},
+
+		/* 	This function is used to bookmark a poll,
+			value is an object data from the poll where we destructure and get the id of that particular poll
+			UserID is the id of the particular loggedIn user which we received from our getters
+			we use these values to update the selected_poll and poll_creator which are keys in our data
+			After the update we dispatch the likePoll action passing in the data as parameter
+		*/
 		likePoll(value, userID) {
 			const { id } = { ...value };
-
 			this.selected_poll = id;
 			this.poll_creator = userID;
 			let data = {
@@ -487,6 +403,13 @@ export default {
 				.dispatch("likePoll", data)
 				.then(() => this.$router.push("/feeds"));
 		},
+
+		/* 	This function is used to bookmark a poll,
+			value is an object data from the poll where we destructure and get the id of that particular poll
+			UserID is the id of the particular loggedIn user which we received from our getters
+			we use these values to update the selected_poll and poll_creator which are keys in our data
+			After the update we dispatch the bookmarkPoll action passing in the data as parameter
+		*/
 		bookmarkPoll(value, userID) {
 			const { id } = { ...value };
 
@@ -500,16 +423,30 @@ export default {
 				.dispatch("bookmarkPoll", data)
 				.then(() => this.$router.push("/"));
 		},
-		...mapActions(["getPolls"])
-	},
-	components: {
-		PollForm,
-		LoginForm,
-		Login,
-		PollMenu,
-		InfiniteLoading
+
+		/* 	This function is used to vote on a choice attached to a poll,
+			value is an object data from the poll where we destructure and get the id of that particular poll
+			UserID is the id of the particular choice which we received when we looped over the choices
+			we use these values to update the selected_poll and selected_choice which are keys in our data
+			After the update we dispatch the voteChoice action passing in the data as parameter
+		*/
+		voteChoice(value, userID) {
+			const { id } = { ...value };
+			this.selected_poll = id;
+			this.selected_choice = userID;
+			let data = {
+				poll: this.selected_poll,
+				choice: this.selected_choice
+			};
+			this.$store
+				.dispatch("voteChoice", data)
+				.then(() => this.$router.push("/"));
+		},
+		// Here we call our getPolls action
+		...mapActions(["getPolls", "getTrendingPolls"])
 	},
 	computed: {
+		// This returns all our updated state
 		...mapGetters([
 			"allPolls",
 			"getUser",
@@ -543,5 +480,6 @@ export default {
 }
 .linkHover:hover {
 	cursor: pointer !important;
+	color: #20aee3;
 }
 </style>